review comments

+++ km/km_management.c
@@ -47,7 +47,9 @@ static void* mgt_main(void* arg)
       km_warnx("Connection accepted");
       close(nfd);
 
-      km_snapshot_create(NULL, 0);
+      if (km_snapshot_create(NULL, 0) != 0) {
+         continue;
+      }
       unlink(addr.sun_path);
       exit(0);
    }
